OpenACR backport Jun-2023
Summary of changes:

New tools introduction:
- AMS - Algo Message System - unified messaging for an enterprise, low latency inter-process comminication framework;
- FM - Dynamic fault management - framework for trouble detection, isolation, and administration;
- ATF_CI - framework for continuous integration;
- ATF_COMP - component test framework;
- ATF_COV - code coverage measurement tool;
- TOAMC - script-helper to port native C/C++ structs to acr/amc;
- FIX-GEN-CONFLICTS - tool for fixing conflicts in generated code.

ATF_CI (former ATF_NORM):
- phases of integration (CI jobs):
- normalization (as in ATF_NORM),
- run tests (including coverage measurement and checking),
- run tests with memory checker;
- automatic capture canned test output for normalization test;
- automatic prepartion of custom environment for tests.

AMC:
- varlen-of-varlen support - now Opt/Varlen field may be in turn of Opt/Varlen ctype;
- AMS stream pnew;
- fix handling of `-verbose` in Execv;
- conditional functions (prepcond);
- check cheap copy on opt/varlen cpptype;
- functions in global namespace;
- generate `WinMain()`;
- move steps from `MainLoop()` to separate `Steps()` function;
- fix error in SetDelay();
- generate more comparison ops: `>`, `<=`, '>=';
- allocate and fill N elements with value - `AllocNVal()` for Tary;
- selectable type for gconst: `const char*`, `strptr`, ...Pkey type, etc.;
- print Tary in JSON format;
- fix `\*_GetOrCreate` NULL return (cppcheck warning);
- fix number parsing;
- minor fixes.

ABT:
- allow ccache in configs other than release;
- fix installed count in report;
- minor improvements.

ACR:
- acr_my: allow input to be a regular file, not dir only.
- acr_ed: simplify message creation;
- acr_ed: create new CI test for ATF_CI;
- acr_ed: create new ssimfile - check for ssimdb namespace;
- acr_compl: fix unique list completion;
- acr_compl: fix UniqueCompletionQ;
- acr_compl: add debuggability.

ALGO_LIB:
- prlog improvement: log category introduction, message timestamp;
- WinChmod644 - set unix-like 644 permissions on Win;
- OpenFile fixes for Win;
- UnTime reading from date only;
- print base32-coded u64;
- kill whole process tree (Linux-only);
- zero-based month format specifier for TimeStruct;
- minor cosmetic changes.

LIB_CTYPE:
- normalize ssim tuple;
- stabilize ssim tuple - mask unstable fields;

ATF:
- atf_ci: stop testing after N errors;
- atf_unit: more tests for number parsing;
- atf_unit: debug path adjustment;
- atf_unit: allow skipping of untracked files check;
- atf_unit: check timezone;
- atf_unit: relax timer test for stability;
- atf_unit: minor fixes;
- unit test enhancement for UnTime printing and parsing;
- unit test for kill whole process tree;
- atf_amc: less verbose test output;
- code optimization and cosmetic changes.

GITLAB:
- work with labels;
- work with milestones;
- automatically assign issue to one who started working on it;
- automatically close issue with `closes #...`.

HILITE:
- move hardcoded color patterns to AMC table;
- extend color markings.

UPDATE-SCRIPTFILE:
- count Python scripts.

+++ cpp/amc/tpool.cpp
@@ -40,7 +40,7 @@ void amc::tclass_Tpool() {
 
     GenTclass(amc_tclass_Pool);
 
-    InsVar(R, field.p_ctype    , "u32", "$name_blocksize", "", "# bytes per block");
+    InsVar(R, field.p_ctype    , "u64", "$name_blocksize", "", "# bytes per block");
     InsVar(R, field.p_ctype    , "$Cpptype*", "$name_free", "", "");
     InsVar(R, field.p_arg      , "$Cpptype*", "$name_next", "", "Pointer to next free element int tpool");
     if (mtfree) {
